Some renaming and refactoring
